+++ src/parsers/choice.c
@@ -160,7 +160,7 @@ HParser *h_choice__mv(HAllocator *mm__, HParser *p, va_list ap_) {
 HParser *h_choice__a(void *args[]) { return h_choice__ma(&system_allocator, args); }
 
 HParser *h_choice__ma(HAllocator *mm__, void *args[]) {
-    size_t len = -1; // because do...while
+    size_t len = (size_t)-1; // because do...while
     const HParser *arg;
 
     do {
